Moved profile button to editor header
